Converting states and next states to be stored as float tensors to accomodate for floating point values

+++ bark_ml/library_wrappers/lib_fqf_iqn_qrdqn/memory/base.py
@@ -95,16 +95,16 @@ class LazyMemory(dict):
   def _sample(self, indices, batch_size):
     bias = -self._p if self._n == self.capacity else 0
 
-    states = np.empty((batch_size, *self.state_shape), dtype=np.uint8)
-    next_states = np.empty((batch_size, *self.state_shape), dtype=np.uint8)
+    states = np.empty((batch_size, *self.state_shape), dtype=np.float32)
+    next_states = np.empty((batch_size, *self.state_shape), dtype=np.float32)
 
     for i, index in enumerate(indices):
       _index = np.mod(index + bias, self.capacity)
       states[i, ...] = self['state'][_index]
       next_states[i, ...] = self['next_state'][_index]
 
-    states = torch.ByteTensor(states).to(self.device).float()
-    next_states = torch.ByteTensor(next_states).to(self.device).float()
+    states = torch.FloatTensor(states).to(self.device)
+    next_states = torch.FloatTensor(next_states).to(self.device)
     actions = torch.LongTensor(self['action'][indices]).to(self.device)
     rewards = torch.FloatTensor(self['reward'][indices]).to(self.device)
     dones = torch.FloatTensor(self['done'][indices]).to(self.device)
